# openapi.yaml
openapi: 3.1.0
info:
  title: Nexora Media Processing API
  description: |
    Plataforma profissional de ingest, transcoding e entrega de media para broadcast e OTT.

    ## Autenticação
    Todos os endpoints requerem JWT RS256. Obter token via `POST /auth/token` (não incluído nesta spec — interno).

    ## ADRs Relevantes
    - **ADR-002**: FFmpeg via execFile/spawn (nunca exec() com string)
    - **ADR-004**: yuv420p obrigatório em outputs de distribuição
    - **ADR-005**: Two-pass EBU R128 + BS1770GAIN
    - **ADR-007**: Audit trail append-only
    - **ADR-010**: VMAF score para todos os outputs
  version: 0.1.0
  contact:
    name: Nexora Engineering
servers:
  - url: http://localhost:3000
    description: Desenvolvimento
  - url: https://api.nexora.io
    description: Produção
security:
  - bearerAuth: []

# ══════════════════════════════════════════════════════════════════
# PATHS
# ══════════════════════════════════════════════════════════════════
paths:
...
  /api/v1/assets/{id}/review:
    post:
      tags: [Assets]
      summary: Aprovar ou rejeitar asset em quarentena
      description: Envia sinal ao workflow Temporal para desbloquear o pipeline.
      parameters:
        - $ref: '#/components/parameters/AssetId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision:
                  type: string
                  enum: [approve, reject]
                reason:
                  type: string
                  maxLength: 1000
      responses:
        '200':
          description: Decisão enviada ao workflow
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:    { type: boolean }
                  assetId:    { type: string }
                  decision:   { type: string }
                  workflowId: { type: string }
                  message:    { type: string }
        '409':
          description: Asset não está em quarentena ou sem workflow activo

  # ── Jobs ─────────────────────────────────────────────────────────
...
  /api/v1/profiles/{id}:
    get:
      tags: [Encoding]
      summary: Obter perfil de encoding por ID
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            enum: [broadcast-hd, ott-hd, web-sd, proxy]
      responses:
        '200':
          description: Perfil encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EncodingProfile'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Metrics ──────────────────────────────────────────────────────
...
  /api/v1/webhooks:
    post:
      tags: [Webhooks]
      summary: Registar webhook para notificações
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebhookRequest'
      responses:
        '201':
          description: Webhook registado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookRegistration'
        '400':
          $ref: '#/components/responses/ValidationError'

    get:
      tags: [Webhooks]
      summary: Listar webhooks activos
      responses:
        '200':
          description: Lista de webhooks
          content:
            application/json:
              schema:
                type: object
                properties:
                  webhooks:
                    type: array
                    items:
                      $ref: '#/components/schemas/WebhookRegistration'
                  count: { type: integer }
  /api/v1/webhooks/{id}:
    get:
      tags: [Webhooks]
      summary: Obter webhook por ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Webhook encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookRegistration'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Webhooks]
      summary: Remover webhook
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Webhook removido
        '404':
          $ref: '#/components/responses/NotFound'

# ══════════════════════════════════════════════════════════════════
# COMPONENTS
# ══════════════════════════════════════════════════════════════════
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT RS256 — assinar com chave privada RSA 4096-bit (ADR-008)
  parameters:
    AssetId:
      name: id
      in: path
      required: true
      description: ID do asset (UUID)
      schema:
        type: string
        format: uuid
  responses:
    ValidationError:
      description: Dados de input inválidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: VALIDATION_ERROR
            message: Dados de input inválidos
...
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: NOT_FOUND
            message: Recurso não encontrado
  schemas:
...
    WebhookEventType:
      type: string
      enum:
        - ASSET_INGESTED
        - ASSET_QC_PASSED
        - ASSET_QC_QUARANTINED
        - ASSET_QC_REJECTED
        - ASSET_TRANSCODING_STARTED
        - ASSET_TRANSCODING_COMPLETED
        - ASSET_PROCESSING_COMPLETED
        - ASSET_PROCESSING_FAILED
        - JOB_COMPLETED
        - JOB_FAILED

    # ── Entidades ───────────────────────────────────────────────────
...
    WebhookRegistration:
      type: object
      properties:
        id:        { type: string, format: uuid }
        url:       { type: string, format: uri }
        events:
          type: array
          items:
            $ref: '#/components/schemas/WebhookEventType'
        isActive:  { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateWebhookRequest:
      type: object
      required: [url, events]
      properties:
        url:
          type: string
          format: uri
          maxLength: 2048
          example: https://meu-sistema.io/webhooks/nexora
        events:
          type: array
          items:
            $ref: '#/components/schemas/WebhookEventType'
          minItems: 1
          example: [ASSET_PROCESSING_COMPLETED, ASSET_PROCESSING_FAILED]
        secret:
          type: string
          maxLength: 256
          description: Segredo HMAC-SHA256 para assinatura do payload
    EncodingProfile:
      type: object
      properties:
        id:          { type: string, example: broadcast-hd }
        name:        { type: string }
        description: { type: string }
        priority:    { type: integer }
        vmafThreshold: { type: integer }
        container:   { type: string }
        useCases:
          type: array
          items: { type: string }
        video:
          type: object
          properties:
            codec:        { type: string }
            bitrateKbps:  { type: integer }
            maxrateKbps:  { type: integer }
            pixelFormat:  { type: string }
            colorspace:   { type: string, nullable: true }
            gopSize:      { type: integer }
            bFrames:      { type: integer }
            closedGop:    { type: boolean }
            cfr:          { type: boolean }
        audio:
          type: object
          properties:
            codec:              { type: string }
            sampleRate:         { type: integer }
            loudnessTargetLufs: { type: number }
            truePeakLimitDbtp:  { type: number }
...
    Error:
      type: object
      properties:
        error:   { type: string }
        message: { type: string }
        details:
          type: object
          nullable: true

# ══════════════════════════════════════════════════════════════════
# TAGS
# ══════════════════════════════════════════════════════════════════
...